Adding new namespace Quickpay::API::Client
