+++ lib/quickpay.rb
@@ -1,10 +1,5 @@
 require 'json'
 require 'httparty'
-
 require 'quickpay/version'
 require 'quickpay/logger'
 require 'quickpay/constants'
-require 'quickpay/errors'
-require 'quickpay/request'
-require 'quickpay/client'
-
