+++ lib/quickpay/api/errors/bad_request.rb
@@ -1,0 +1,7 @@
+
+module QuickPay
+  module API
+    class BadRequest < Error
+    end
+  end
+end
