+++ lib/quickpay/api/errors/conflict.rb
@@ -1,0 +1,7 @@
+
+module QuickPay
+  module API
+    class Conflict < Error
+    end
+  end
+end
