+++ lib/quickpay/api/errors/forbidden.rb
@@ -1,0 +1,7 @@
+
+module QuickPay
+  module API
+    class Forbidden < Error
+    end
+  end
+end
