+++ lib/quickpay/api/errors/method_not_allowed.rb
@@ -1,0 +1,7 @@
+
+module QuickPay
+  module API
+    class MethodNotAllowed < Error
+    end
+  end
+end
