+++ lib/quickpay/api/errors/not_acceptable.rb
@@ -1,0 +1,7 @@
+
+module QuickPay
+  module API
+    class NotAcceptable < Error
+    end
+  end
+end
