+++ lib/quickpay/api/errors/not_found.rb
@@ -1,0 +1,7 @@
+
+module QuickPay
+  module API
+    class NotFound < Error
+    end
+  end
+end
