+++ lib/quickpay/api/errors/payment_required.rb
@@ -1,0 +1,7 @@
+
+module QuickPay
+  module API
+    class PaymentRequired < Error
+    end
+  end
+end
