+++ lib/quickpay/api/errors/quickpay_error.rb
@@ -1,0 +1,13 @@
+module QuickPay
+  module API
+    class Error < StandardError
+      attr_reader :code, :status, :body
+
+      def initialize(code, status, body)
+        @code   = code
+        @status = status
+        @body   = body
+      end
+    end
+  end  
+end
