+++ lib/quickpay/api/errors/server_error.rb
@@ -1,0 +1,7 @@
+
+module QuickPay
+  module API
+    class ServerError < Error
+    end
+  end
+end
