+++ lib/quickpay/api/errors/unauthorized.rb
@@ -1,0 +1,7 @@
+
+module QuickPay
+  module API
+    class Unauthorized < Error
+    end
+  end
+end
