--- lib/quickpay/errors/bad_request.rb
@@ -1,6 +1,0 @@
-
-module QuickPay
-  class BadRequest < Error
-      
-  end
-end
