--- lib/quickpay/errors/conflict.rb
@@ -1,6 +1,0 @@
-
-module QuickPay
-  class Conflict < Error
-      
-  end
-end
