--- lib/quickpay/errors/forbidden.rb
@@ -1,6 +1,0 @@
-
-module QuickPay
-  class Forbidden < Error
-      
-  end
-end
