--- lib/quickpay/errors/method_not_allowed.rb
@@ -1,6 +1,0 @@
-
-module QuickPay
-  class MethodNotAllowed < Error
-      
-  end
-end
