--- lib/quickpay/errors/not_acceptable.rb
@@ -1,6 +1,0 @@
-
-module QuickPay
-  class NotAcceptable < Error
-      
-  end
-end
