--- lib/quickpay/errors/not_found.rb
@@ -1,6 +1,0 @@
-
-module QuickPay
-  class NotFound < Error
-      
-  end
-end
