--- lib/quickpay/errors/payment_required.rb
@@ -1,6 +1,0 @@
-
-module QuickPay
-  class PaymentRequired < Error
-      
-  end
-end
