--- lib/quickpay/errors/quickpay_error.rb
@@ -1,13 +1,0 @@
-module QuickPay
-
-  class Error < StandardError
-    attr_reader :code, :status, :body
-
-    def initialize(code, status, body)
-      @code   = code
-      @status = status
-      @body   = body
-    end
-  end
-  
-end
