--- lib/quickpay/errors/server_error.rb
@@ -1,6 +1,0 @@
-
-module QuickPay
-  class ServerError < Error
-      
-  end
-end
