--- lib/quickpay/errors/unauthorized.rb
@@ -1,6 +1,0 @@
-
-module QuickPay
-  class Unauthorized < Error
-      
-  end
-end
