+++ spec/spec_helper.rb
@@ -2,7 +2,7 @@ $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 
 require 'bundler/setup'
 Bundler.setup(:default, :development)
-require 'quickpay'
+require 'quickpay/api/client'
 require 'webmock/rspec'
 
 QuickPay.logger.level = Logger::DEBUG
